% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alta_lake}
\alias{alta_lake}
\title{Alta Lake Gravity Core Data}
\format{A \link{mudata} object}
\usage{
alta_lake
}
\description{
Bulk geochemistry of a gravity core from Alta Lake, Whistler, British Columbia,
Canada.
}
\examples{
library(ggplot2)
autoplot(alta_lake, y = "depth") + scale_y_reverse()
autoplot(alta_lake, y = "age")

}
\references{
Dunnington DW, Spooner IS, White CE, et al (2016) A geochemical perspective on the impact of 
development at Alta Lake, British Columbia, Canada. J Paleolimnol 56:315-330. 
doi: 10.1007/s10933-016-9919-x
}
\keyword{datasets}
