% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{as_mudata}
\alias{as_mudata}
\alias{as.mudata}
\alias{as_mudata.mudata}
\alias{as_mudata.data.frame}
\alias{as_mudata.tbl}
\alias{as_mudata.list}
\title{Coerce objects to mudata}
\usage{
as_mudata(x, ...)

as.mudata(x, ...)

\method{as_mudata}{mudata}(x, ...)

\method{as_mudata}{data.frame}(x, ...)

\method{as_mudata}{tbl}(x, ...)

\method{as_mudata}{list}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Passed to other methods}
}
\value{
A \link{mudata} object or an error
}
\description{
Coerce objects to mudata
}
