% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.qtag.long}
\alias{autoplot.qtag.long}
\alias{autoplot.qtag.wide}
\alias{plot.qtag.long}
\alias{plot.qtag.wide}
\title{Autoplot a qualifier/tag structure}
\usage{
\method{autoplot}{qtag.long}(x, subset, xvar, yvar, facets, geom = "path",
  errors = "err", ...)

\method{autoplot}{qtag.wide}(x, ...)

\method{plot}{qtag.long}(x, ...)

\method{plot}{qtag.wide}(x, ...)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{subset}{Subset to plot}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{facets}{Column to be used as facetting variable}

\item{geom}{GGPlot geometries to be used. Can be any combination of point, path, or line.}

\item{errors}{The colum that contains uncertainty information}

\item{...}{Passed on to \code{aes_string()}}
}
\value{
A ggplot object
}
\description{
Autoplot a qualifier/tag structure
}
\examples{
data(pocmaj)

pocmajqt <- as.qtag(pocmaj, .qualifiers=c("core", "depth"))
plot(pocmajqt, geom=c("path", "point"))
plot(pocmajqt, subset=core=="MAJ-1" & column \%in\% c("Ca", "Ti"))
plot(pocmajqt, shape="core")
plot(long(pocmajqt))

library(ggplot2)
autoplot(pocmajqt, shape="core")

}

