% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot.qtag.long}
\alias{biplot.qtag.long}
\alias{biplot.qtag.wide}
\title{Biplot an object using ggplot}
\usage{
\method{biplot}{qtag.long}(x, namesx = NULL, namesy = NULL, values = NULL,
  namecolumn = NULL, errors = NULL, labeller = ggplot2::label_value, ...)

\method{biplot}{qtag.wide}(x, ...)
}
\arguments{
\item{x}{the object to biplot}

\item{namesx}{The names to be included in the x axes, or all the names to be included}

\item{namesy}{The names to be included on the y axes, or NULL for all possible combinations
of \code{namesx}.}

\item{values}{The column containing the values to plot}

\item{namecolumn}{The column where namesx and namesy are to be found}

\item{errors}{The column containing the errors. Use \code{NULL} for default ("err" if column
exists or none otherwise), or \code{NA} to suppress.}

\item{labeller}{The labeller to use to label facets (may want to use \code{label_parsed}
to use plotmath-style labels)}

\item{...}{passed to \code{aes_string()}}
}
\description{
Biplot an object using ggplot
}
\examples{
data(pocmaj)
qt <- as.qtag(pocmaj)
biplot(qt, color="core")

}

