% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{biplotgg.mudata}
\alias{biplotgg.mudata}
\title{Biplot a mudata object}
\usage{
\method{biplotgg}{mudata}(x, ...)
}
\arguments{
\item{x}{A mudata object}

\item{...}{passed to \link{biplotgg.qtag.long}}
}
\value{
a ggplot object
}
\description{
Biplot a mudata object
}

