% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplotgg}
\alias{biplotgg}
\alias{biplotgg.qtag.long}
\title{Biplot an object using ggplot}
\usage{
biplotgg(x, ...)

\method{biplotgg}{qtag.long}(x, column, namesx, namesy = NULL,
  values = NULL, errors = NULL, labeller = ggplot2::label_value, ...)
}
\arguments{
\item{x}{the object to biplot}

\item{...}{passed to other methods}
}
\description{
Biplot an object using ggplot
}

