% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{collect.mudata}
\alias{collect.mudata}
\title{Collect all mudata components}
\usage{
\method{collect}{mudata}(x, ...)
}
\arguments{
\item{x}{A mudata object}

\item{...}{Passed to \link[dplyr]{collect}}
}
\value{
A mudata object with all components as local data frames.
}
\description{
Objects created by \link{mudata} are generally assumed to be local data frames,
but some methods may function on database tbls (especially in the future). 
This function applies \link[dplyr]{collect} to all component tables.
}
