% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{condense.tags}
\alias{condense.tags}
\alias{condense.tags.data.frame}
\alias{condense.tags.mudata}
\title{Condense multiple columns to a single JSON column}
\usage{
condense.tags(x, tagcolumns, tagcolumn)

\method{condense.tags}{data.frame}(df, tagcolumns, tagcolumn = "tags")

\method{condense.tags}{mudata}(x, ...)
}
\arguments{
\item{tagcolumns}{column names to be condensed to JSON}

\item{df}{a data frame with tag columns}
}
\value{
A modified data.frame
}
\description{
Condense multiple columns to a single JSON column
}

