% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata_helpers.R
\name{data_tbl}
\alias{data_tbl}
\alias{params_tbl}
\alias{locations_tbl}
\alias{datasets_tbl}
\alias{columns_tbl}
\alias{x_columns}
\title{Access components of a mudata object}
\usage{
data_tbl(x)

params_tbl(x)

locations_tbl(x)

datasets_tbl(x)

columns_tbl(x)

x_columns(x)
}
\arguments{
\item{x}{A mudata object}
}
\value{
The appropriate component
}
\description{
Access components of a mudata object
}
\examples{
data_tbl(kentvillegreenwood)

}
