% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{expand.tags}
\alias{expand.tags}
\title{Expand JSON to multiple columns}
\usage{
expand.tags(x, tagcolumn = "tags", ...)
}
\arguments{
\item{x}{a vector of JSON values or a data.frame with a 'tags' column}

\item{tagcolumn}{the column comtaining json}
}
\value{
A data.frame with columns added
}
\description{
Expand JSON to multiple columns
}

