% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{filter.mudata}
\alias{filter.mudata}
\alias{subset.mudata}
\title{Subset (filter) a MuData object}
\usage{
filter.mudata(x, ..., datasets = NULL, params = NULL, locations = NULL)

\method{subset}{mudata}(x, ..., datasets = NULL, params = NULL,
  locations = NULL)
}
\arguments{
\item{x}{The object to subset}

\item{...}{Aguments to/from methods}

\item{datasets}{Vector of datasets to include}

\item{params}{Vector of parameters to include}

\item{locations}{Vector of locations to include}
}
\value{
A filtered MuData object
}
\description{
Subset (filter) a MuData object
}
