% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{fortify.qtag}
\alias{fortify.qtag}
\title{Fortify a qualifier/tag structure for use in ggplot}
\usage{
\method{fortify}{qtag}(x, skip_aggregate = FALSE, ...)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{skip_aggregate}{Pass \code{TRUE} to skip aggregation}

\item{...}{Passed on to \code{aggregate} (or \link{long} if aggregate is \code{FALSE})}
}
\value{
A \code{qtag.long} object
}
\description{
Fortify a qualifier/tag structure for use in ggplot
}
\examples{
data(pocmaj)
library(ggplot2)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
fortify(pocmaj)
ggplot(pocmaj, aes(x=values, y=depth, col=core)) + geom_point() +
 scale_y_reverse() + facet_wrap(~column)

}

