% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{generate_type_tbl}
\alias{generate_type_tbl}
\alias{generate_type_tbl.mudata}
\alias{generate_type_tbl.tbl}
\alias{generate_type_tbl.data.frame}
\alias{generate_type_str}
\title{Generate type descriptors}
\usage{
generate_type_tbl(x, default = "guess")

\method{generate_type_tbl}{mudata}(x, default = "guess")

\method{generate_type_tbl}{tbl}(x, default = "guess")

\method{generate_type_tbl}{data.frame}(x, default = "guess")

generate_type_str(x, default = "guess")
}
\arguments{
\item{x}{An object}

\item{default}{The default type descriptor (should usually be 'guess')}
}
\value{
A type descriptor or tibble of type descriptors
}
\description{
These functions generate type descriptors for use in the \link{mudata}
columns table. They are used in \link{mudata_parse_column} to deserialize
a vector read from csv or json.
}
\keyword{internal}
