% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{id.vars}
\alias{id.vars}
\alias{id.vars.data.frame}
\alias{id.vars.grouped_df}
\alias{id.vars.qtag}
\title{Extract value qualifier names from a qualifier/tag structure}
\usage{
id.vars(x, ...)

\method{id.vars}{qtag}(x, ...)

\method{id.vars}{grouped_df}(x, ...)

\method{id.vars}{data.frame}(x, ..., quiet = FALSE)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{...}{Passed to/from methods}

\item{quiet}{Supress error messages on coersion to a qualifier/tag structure}
}
\value{
A vector of qualifier column names
}
\description{
Extract value qualifier names from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, id.vars=c("core", "depth"))
id.vars(pocmaj)

}

