% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{is_mudata}
\alias{is_mudata}
\alias{is.mudata}
\title{Test if an object is a mudata object}
\usage{
is_mudata(x)

is.mudata(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object is a mudata object, FALSE otherwise
}
\description{
Test if an object is a mudata object
}
\examples{
is_mudata(kentvillegreenwood)

}
