% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kentvillegreenwood}
\alias{kentvillegreenwood}
\title{Kentville/Greenwood Climate Data}
\format{A \link{mudata} object}
\source{
Environment Canada via the 'rclimateca' package. \url{http://climate.weather.gc.ca/}
}
\usage{
kentvillegreenwood
}
\description{
Climate data for Kentville and Greenwood (Nova Scotia) for July and August of 1999.
}
\examples{
autoplot(kentvillegreenwood)

}
\keyword{datasets}
