% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{long}
\alias{long}
\alias{long.default}
\alias{long.qtag.long}
\alias{long.qtag.wide}
\title{Convert data to long format}
\usage{
long(x, ...)

\method{long}{default}(x, ...)

\method{long}{qtag.long}(x, ...)

\method{long}{qtag.wide}(qtag, varname = "column", quiet = FALSE)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{...}{Passed to other methods}
}
\value{
A (possibly unchanged) \code{qtag.long} data.frame
}
\description{
Convert data to long format
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers = c("core", "depth"))
long(pocmaj)

}

