% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{long_ggplot}
\alias{long_ggplot}
\alias{long_plot}
\alias{long_plot_base}
\title{Smart plotting of tidy data frames}
\usage{
long_ggplot(.data, ..., facet_args = list())

long_plot(.data, ..., max_facets = 9, facet_args = list(),
  scales = list())

long_plot_base(.data, id_vars = NULL, measure_var = "value", x = NULL,
  y = NULL, facets = NULL, geom = "path", error_var = NULL, ...)
}
\arguments{
\item{...}{Passed on to \code{aes_string()}}

\item{x}{A \code{data.frame}}

\item{facets}{Column to be used as facetting variable}

\item{geom}{GGPlot geometries to be used. Can be any combination of point, path, or line.}

\item{id.vars}{Columns that identify unique values}

\item{measure.var}{Column that contains values to be plotted}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{subset}{Subset to plot}

\item{errors}{The colum that contains uncertainty information}
}
\value{
A ggplot object
}
\description{
This function uses ggplot to plot a 'long' data frame with a few \code{id.vars}, 
or variables that identify the values in the \code{value} column. The function
is optimised to plot multi-parameter spatiotemporal data either horizontally
(time on the x axis) or vertically (time on the y axis). Facets are intended
to be by parameter, which is guessed based on the right-most variable named
in \code{id.vars}. This is intended to produce a quick visual of an object to
examine its contents.
}
\examples{
data(pocmajsum)
long_plot(pocmajsum, c("core", "depth"), "Ca")

library(tidyr)
pocmaj_long <- pocmajsum \%>\%
  gather(Ca, Ti, V, key = "variable", value = "value")
long_plot(pocmaj_long, col="core")
long_ggplot(pocmaj_long, col="core")

}
