% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longlake2016}
\alias{longlake2016}
\title{Sample mudata object}
\format{A \code{mudata} object containing multi-parameter spatiotemporal data}
\usage{
longlake2016
}
\description{
Example of a sample mudata object
}
\keyword{datasets}

