% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{measure.vars}
\alias{measure.vars}
\title{Extract value column names from a qualifier/tag structure}
\usage{
measure.vars(x, quiet = FALSE)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{quiet}{Suppress error messages on coersion to a qualifier/tag structure}
}
\value{
A vector of value column names
}
\description{
Extract value column names from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, id.vars=c("core", "depth"))
measure.vars(pocmaj)

}

