% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{mudata}
\alias{mudata}
\title{Create a MuData object}
\usage{
mudata(data, locations = NULL, params = NULL, datasets = NULL,
  columns = NULL, dataset.id = "default", location.id = "default",
  defactorize = TRUE, validate = TRUE, expand.tags = TRUE,
  retype = TRUE)
}
\arguments{
\item{data}{The data table}

\item{locations}{The locations table (can be omitted)}

\item{params}{The params table (can be omitted)}

\item{datasets}{The datasets table (can be omitted)}

\item{columns}{The columns table (can be omitted)}

\item{dataset.id}{The dataset id (if datasets is unspecified)}

\item{location.id}{The location id (if locations is unspecified)}

\item{defactorize}{Pass \code{FALSE} to keep input columns as factors (may cause errors).}

\item{validate}{Flag to validate input}

\item{expand.tags}{Flag to expand JSON tags to columns}

\item{retype}{Pass \code{TRUE} to retype columns based on the 'type' column of the 'columns'
table.}
}
\value{
A \code{mudata} object
}
\description{
Create a MuData object
}
\examples{
data(pocmaj)
# melt and aggregate data
pocmaj <- as.qtag(pocmaj)
pocmaj <- long(pocmaj)
pocmaj <- aggregate(pocmaj)
# rename columns so there is a location, param, x, and value
datatable <- rename.cols(pocmaj, core="location", depth="x")
md <- mudata(datatable)
summary(md)
plot(md, yvar="x", geom=c("path", "point"))

}

