% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata2-package.R
\docType{package}
\name{mudata2-package}
\alias{mudata2}
\alias{mudata2-package}
\title{mudata2: Interchange Tools for Multi-Parameter Spatiotemporal Data}
\description{
Formatting and structuring multi-parameter spatiotemporal data is often a time-consuming task. This package offers functions and data structures designed to easily organize and visualize these data for applications in geology, paleolimnology, dendrochronology, and paleoclimate. See Dunnington and Spooner (2018) \doi{10.1139/facets-2017-0026}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/mudata2/}
  \item \url{https://github.com/paleolimbot/mudata2}
  \item Report bugs at \url{https://github.com/paleolimbot/mudata2/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
