% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{mudata_sql}
\alias{mudata_sql}
\title{Create a mudata object using a database source}
\usage{
mudata_sql(db, data = "data", locations = NA, params = NA,
  datasets = NA, columns = NA)
}
\arguments{
\item{db}{An src_sql as generated by dplyr}

\item{data}{The name of the data table}

\item{locations}{The name of the locations table}

\item{params}{The name of the params table}

\item{datasets}{The name of the datasets table}

\item{columns}{The name of the columns table}
}
\value{
A mudata object
}
\description{
Create a mudata object using a database source
}
\keyword{internal}
