% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_climate}
\alias{ns_climate}
\title{Nova Scotia Long-Term Climate Data}
\format{
A \link{mudata} object
}
\source{
Environment Canada: \url{https://climate.weather.gc.ca/}
}
\usage{
ns_climate
}
\description{
Monthly climate data for locations in Nova Scotia with records longer than 80 years.
}
\examples{
print(ns_climate)

}
\keyword{datasets}
