% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel_melt}
\alias{parallel_melt}
\alias{parallel_gather}
\alias{parallel_gather_}
\title{Melt multiple sets of columns in parallel}
\usage{
parallel_melt(x, id.vars, ..., variable.name = "param",
  factorsAsStrings = TRUE)

parallel_gather(x, key, ..., convert = FALSE, factor_key = FALSE)

parallel_gather_(x, key, ..., convert = FALSE, factor_key = FALSE)
}
\arguments{
\item{x}{A data.frame}

\item{id.vars}{vector of ID variable names}

\item{...}{Named arguments specifying the \code{measure.vars} to be stored to the
column name specified.}

\item{variable.name, key}{Column name to use to store variables}

\item{factorsAsStrings, factor_key}{Control whether factors are converted to character when melted as
measure variables.}

\item{convert}{Convert types (see \link[tidyr]{gather})}
}
\value{
A molten data.frame
}
\description{
Essentially this is a wrapper around \link[reshape2]{melt.data.frame} that
is able to \link{cbind} several melt operations. This is useful when a wide
data frame contains uncertainty or flag information in paired columns.
}
\examples{
data(pocmajsum)
parallel_melt(pocmajsum,
              id.vars=c("core", "depth"),
              values=c("Ca", "Ti", "V"),
              err=c("Ca_sd", "Ti_sd", "V_sd"))

}
