% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{plotgg.mudata}
\alias{plotgg.mudata}
\title{Autoplot a mudata object}
\usage{
\method{plotgg}{mudata}(x, ...)
}
\arguments{
\item{x}{A \link{mudata} object}

\item{...}{Passed on to \code{plotgg.qtag.long}}
}
\value{
A ggplot object
}
\description{
Autoplot a mudata object
}

