% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_qtag.R
\name{plotgg.qtag.long}
\alias{plotgg}
\alias{plotgg.qtag.long}
\alias{plotgg.qtag.wide}
\title{Autoplot a qualifier/tag structure}
\usage{
\method{plotgg}{qtag.long}(x, subset, xvar, yvar, facets, errors = "err", ...)

\method{plotgg}{qtag.wide}(x, ...)

plotgg(x, ...)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{subset}{Subset to plot}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{facets}{Column to be used as facetting variable}

\item{errors}{The colum that contains uncertainty information}

\item{...}{Passed on to \code{aes_string()}}
}
\value{
A ggplot object
}
\description{
Autoplot a qualifier/tag structure
}
\examples{
data(pocmaj)
plotgg(pocmaj)

pocmajqt <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
plotgg(pocmajqt)
plotgg(pocmajqt, subset=core=="MAJ-1" & column \%in\% c("Ca", "Ti"))
plotgg(pocmajqt, shape="core")


}

