% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pocmaj}
\alias{pocmaj}
\title{Pockwock Lake/Lake Major Elemental Sample Data}
\format{
A \code{data.frame} containing multi-qualifier concentration data
}
\usage{
pocmaj
}
\description{
A small example data.frame used to test structure methods.
}
\keyword{datasets}
