% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pocmajsum}
\alias{pocmajsum}
\title{Pre-summarised Sample Data}
\format{
A \code{data.frame} containing multi-qualifier data
}
\usage{
pocmajsum
}
\description{
A small example data.frame of pre-summarised data; a summarised
version of the \link{pocmaj} dataset.
}
\keyword{datasets}
