% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{print.mudata}
\alias{print.mudata}
\alias{summary.mudata}
\title{Print a mudata object}
\usage{
\method{print}{mudata}(x, ..., width = NULL)

\method{summary}{mudata}(object, ...)
}
\arguments{
\item{x, object}{A mudata object}

\item{...}{Passed to other methods}

\item{width}{The number of characters to use as console width}
}
\value{
print returns x (invisibly); summary returns a data frame with summary information.
}
\description{
Print a mudata object
}
\examples{
print(kentvillegreenwood)
summary(kentvillegreenwood)

}
