% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{qualifierdata}
\alias{qualifierdata}
\title{Extract qualifier column data from a qualifier/tag structure}
\usage{
qualifierdata(x)
}
\arguments{
\item{x}{A \link{qtag} object}
}
\value{
A \link{data.frame} of qualifier column data
}
\description{
Extract qualifier column data from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
qualifierdata(pocmaj)

}

