% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{qualifierplot}
\alias{qualifierplot}
\title{Smart plotting of tidy data frames}
\usage{
qualifierplot(x, id.vars, measure.var, subset, xvar, yvar, facets,
  geom = "path", errors = "err", ...)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{id.vars}{Columns that identify unique values}

\item{measure.var}{Column that contains values to be plotted}

\item{subset}{Subset to plot}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{facets}{Column to be used as facetting variable}

\item{geom}{GGPlot geometries to be used. Can be any combination of point, path, or line.}

\item{errors}{The colum that contains uncertainty information}

\item{...}{Passed on to \code{aes_string()}}
}
\value{
A ggplot object
}
\description{
This function uses ggplot to plot a 'long' data frame with a few \code{id.vars}, 
or variables that identify the values in the \code{value} column. The function
is optimised to plot multi-parameter spatiotemporal data either horizontally
(time on the x axis) or vertically (time on the y axis). Facets are intended
to be by parameter, which is guessed based on the right-most variable named
in \code{id.vars}. This is intended to produce a quick visual of an object to
examine its contents.
}
\examples{
data(pocmajsum)
qualifierplot(pocmajsum, c("core", "depth"), "Ca")

library(reshape2)
pocmaj_long <- melt(pocmajsum, id.vars = c("core", "depth"),
                    measure.vars = c("Ca", "Ti", "V"))
qualifierplot(pocmaj_long, c("core", "depth", "variable"), "value", col="core")


}
