% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{rbind.mudata}
\alias{rbind.mudata}
\title{Combine mudata objects}
\usage{
rbind.mudata(..., validate = TRUE)
}
\arguments{
\item{...}{Mudata objects to combine}

\item{validate}{Flag to validate final object}
}
\value{
A Mudata object
}
\description{
Combine mudata objects
}

