% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{rbind.qtag.long}
\alias{rbind.qtag}
\alias{rbind.qtag.long}
\title{Combine qualifier/tag structures}
\usage{
rbind.qtag.long(...)

rbind.qtag(...)
}
\arguments{
\item{...}{Objects to combine}
}
\value{
A \code{qtag.long} object
}
\description{
Combine qualifier/tag structures
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
newrow <- data.frame(core="POC-2", depth=6, Ca=2100, Ti=4100, V=45)
rbind(pocmaj, newrow)

}

