% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename.cols}
\alias{rename.cols}
\alias{rename.cols.default}
\alias{rename.cols.qtag}
\title{Replace/rename a column in an object}
\usage{
rename.cols(x, ...)

\method{rename.cols}{default}(x, ...)

\method{rename.cols}{qtag}(x, ...)
}
\arguments{
\item{x}{An object that has columns that can be renamed}

\item{...}{Key/value pairs to replace in the form \code{oldval="newval"}}
}
\value{
A copy of the modified object
}
\description{
Essentially a thin convenience wrapper around \code{plyr::rename(x, list(...))},
except \link{qtag} objects have their qualifiers/tags/values attributes properly modified
}
\examples{
data(pocmaj)
rename.cols(pocmaj, Ca="Calcium")
pocmaj2 <- as.qtag(pocmaj, .qualifiers=c("core", "depth"))
pocmaj2 <- rename.cols(pocmaj2, Ca="Calcium")
attr(pocmaj2, "values")

}

