% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename.datasets}
\alias{rename.cols.mudata}
\alias{rename.datasets}
\alias{rename.locations}
\alias{rename.params}
\title{Rename datasets, params, locations, and columns}
\usage{
rename.datasets(md, ..., apply_to = c("data", "locations", "params",
  "datasets", "columns"), warn_missing = TRUE)

rename.params(md, ..., apply_to = c("data", "params"), warn_missing = TRUE)

rename.locations(md, ..., apply_to = c("data", "locations"),
  warn_missing = TRUE)

\method{rename.cols}{mudata}(x, ..., apply_to = c("datasets", "locations",
  "params", "data", "columns"), warn_missing = FALSE,
  warn_duplicated = TRUE)
}
\arguments{
\item{md}{A \link{mudata} object}

\item{...}{Key/value pairs in the form \code{"oldvalue"="newvalue"}}

\item{apply_to}{The sub-elements which the rename operation should consider}

\item{warn_missing}{Print a message if any old names are not actually present}

\item{x}{A \link{mudata} object}

\item{warn_duplicated}{Print a message if any name appears more than once in x 
after the operation.}
}
\value{
A modified \link{mudata} object.
}
\description{
Provides a convenient way to rename datasets, params, locations, and columns
such that their usage remains consistent.
}
\examples{
data(longlake2016)
md2 <- rename.datasets(longlake2016, Dunnington_longlake="dll")
validate.mudata(md2)
md2 <- rename.locations(longlake2016, "LL GC10"="LLGC10")
validate.mudata(md2)
md2 <- rename.params(longlake2016, LOI="Loss On Ignition")
validate.mudata(md2)
md2 <- rename.cols(longlake2016, err="stderr")
validate.mudata(md2)

}

