% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_cols_base}
\alias{rename_cols_base}
\alias{rename_cols_base.default}
\title{Rename a column in an object}
\usage{
rename_cols_base(.data, ..., warn_missing = TRUE, warn_duplicated = TRUE)

\method{rename_cols_base}{default}(.data, ..., warn_missing = TRUE, warn_duplicated = TRUE)
}
\arguments{
\item{.data}{An object that has columns that can be renamed}

\item{...}{Key/value pairs to replace in the form \code{oldval="newval"}}

\item{warn_missing}{Print a message if any old names are not actually present in x}

\item{warn_duplicated}{Print a message if any name appears more than once in x
after the operation.}
}
\value{
A copy of the modified object
}
\description{
Rename columns in a data frame or list
}
\keyword{internal}
