% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_locations}
\alias{rename_locations}
\alias{rename_params}
\alias{rename_datasets}
\alias{rename_columns}
\title{Rename things in a mudata object}
\usage{
rename_locations(.data, ...)

rename_params(.data, ...)

rename_datasets(.data, ...)

rename_columns(.data, ...)
}
\arguments{
\item{.data}{A mudata object}

\item{...}{Variables to rename in the form new_var = old_var}
}
\value{
A modified mudata object
}
\description{
Rename things in a mudata object
}
\examples{
rename_datasets(kentvillegreenwood, avalley = ecclimate)
rename_locations(kentvillegreenwood, Greenwood = starts_with("GREENWOOD"))
rename_params(kentvillegreenwood, max_temp = maxtemp)
rename_columns(kentvillegreenwood, lon = longitude, lat = latitude)

}
