% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R, R/mudata_subset.R
\name{collect.mudata}
\alias{collect.mudata}
\alias{subset.mudata}
\alias{select_datasets}
\alias{select_datasets.default}
\alias{select_locations}
\alias{select_locations.default}
\alias{select_params}
\alias{select_params}
\alias{filter_datasets}
\alias{filter_datasets.default}
\alias{filter_data}
\alias{filter_data.default}
\alias{filter_locations}
\alias{filter_locations.default}
\alias{filter_params}
\alias{filter_params}
\title{Subset (filter) a MuData object}
\usage{
\method{collect}{mudata}(x, ...)

\method{subset}{mudata}(x, ..., datasets = NULL, params = NULL,
  locations = NULL)

select_datasets(.data, ...)

\method{select_datasets}{default}(.data, ..., .factor = FALSE)

select_locations(.data, ...)

\method{select_locations}{default}(.data, ..., .factor = FALSE)

select_params(.data, ..., .factor = FALSE)

select_params(.data, ..., .factor = FALSE)

filter_datasets(.data, ...)

\method{filter_datasets}{default}(.data, ...)

filter_data(.data, ...)

\method{filter_data}{default}(.data, ...)

filter_locations(.data, ...)

\method{filter_locations}{default}(.data, ...)

filter_params(.data, ...)

filter_params(.data, ...)
}
\arguments{
\item{x, .data}{The object to subset}

\item{...}{Aguments to/from methods}

\item{datasets}{Vector of datasets to include}

\item{params}{Vector of parameters to include}

\item{locations}{Vector of locations to include}

\item{.factor}{Maintain order of selected components by converting them to factors}
}
\value{
A subsetted mudata object
}
\description{
Subset (filter) a MuData object
}
