% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{summary.mudata}
\alias{summary.mudata}
\title{Object summary for a mudata object}
\usage{
\method{summary}{mudata}(object, ..., digits = NA)
}
\arguments{
\item{object}{A \link{mudata} object}

\item{...}{Unused}

\item{digits}{The number of digits to be used for rounding, or NA to suppress rounding.}
}
\value{
A data.frame containing the summary
}
\description{
Returns a data.frame containing summary (by dataset, location, and param)
statistics.
}
\examples{
data(longlake2016)
summary(longlake2016, digits=NA)
summary(longlake2016)

}

