% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{tagdata}
\alias{tagdata}
\title{Extract tag column data from a qualifier/tag structure}
\usage{
tagdata(x)
}
\arguments{
\item{x}{A \link{qtag} object}
}
\value{
A \link{data.frame} of tag column data
}
\description{
Extract tag column data from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
tagdata(pocmaj)

}

