% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{tags}
\alias{tags}
\title{Extract tag column names from a qualifier/tag structure}
\usage{
tags(x)
}
\arguments{
\item{x}{A \link{qtag} object}
}
\value{
A vector of tag column names
}
\description{
Extract tag column names from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
tags(pocmaj)

}

