% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata_helpers.R
\name{tbl_data}
\alias{tbl_data}
\alias{tbl_params}
\alias{tbl_locations}
\alias{tbl_datasets}
\alias{tbl_columns}
\alias{tbl.mudata}
\alias{x_columns}
\title{Access components of a mudata object}
\usage{
tbl_data(x)

tbl_params(x)

tbl_locations(x)

tbl_datasets(x)

tbl_columns(x)

\method{tbl}{mudata}(src, which, ...)

x_columns(x)
}
\arguments{
\item{x, src}{A mudata object}

\item{which}{Which tbl to extract}

\item{...}{Unused}
}
\value{
The appropriate component
}
\description{
Access components of a mudata object
}
\examples{
tbl_data(kentvillegreenwood)

}
