% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata-io.R
\name{update_columns_table}
\alias{update_columns_table}
\title{Update the columns table}
\usage{
update_columns_table(md, quiet = FALSE)
}
\arguments{
\item{md}{A mudata object}

\item{quiet}{Suppress changes to existing types}
}
\value{
A mudata object
}
\description{
Update the columns table
}
