% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata_helpers.R
\name{update_datasets}
\alias{update_datasets}
\alias{update_locations}
\alias{update_params}
\alias{update_columns}
\title{Add documentation to mudata objects}
\usage{
update_datasets(x, datasets, ...)

update_locations(x, locations, datasets, ...)

update_params(x, params, datasets, ...)

update_columns(x, columns, tables, datasets, ...)
}
\arguments{
\item{x}{A mudata object}

\item{datasets}{One or more datasets to update}

\item{...}{Key/value pairs (values of length 1)}

\item{locations}{One or more locations to update}

\item{params}{One or more params to update}

\item{columns}{One or more columns to update (columns table)}

\item{tables}{One or more tables to update (columns table)}
}
\value{
A modified version of x
}
\description{
Add documentation to mudata objects
}
\examples{
kentvillegreenwood \%>\%
  update_datasets("ecclimate", new_key = "new_value") \%>\%
  tbl_datasets()

}
