% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{validate.mudata}
\alias{validate.mudata}
\title{Validate a MUData object}
\usage{
validate.mudata(md)
}
\arguments{
\item{md}{An object of class 'mudata'}
}
\value{
TRUE, invisibly
}
\description{
Validates a MUData object by calling \code{stop} when an error is found.
}
\examples{
data(kentvillegreenwood)
validate.mudata(kentvillegreenwood)

}

