% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{valuedata}
\alias{valuedata}
\title{Extract value column data from a qualifier/tag structure}
\usage{
valuedata(x)
}
\arguments{
\item{x}{A \link{qtag} object}
}
\value{
A \link{data.frame} of value column data
}
\description{
Extract value column data from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
valuedata(pocmaj)

}

