% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{values}
\alias{values}
\title{Extract value column names from a qualifier/tag structure}
\usage{
values(x, quiet = FALSE)
}
\arguments{
\item{x}{A \link{qtag} object}
}
\value{
A vector of value column names
}
\description{
Extract value column names from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
values(pocmaj)

}

