% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{write.mudata}
\alias{read.mudata}
\alias{write.mudata}
\title{Read/Write a MuData zip file}
\usage{
write.mudata(md, zipfile, overwrite = FALSE, expand.tags = TRUE, ...)

read.mudata(zipfile, validate = TRUE, expand.tags = TRUE, load = c("data",
  "locations", "params", "datasets", "columns"), ...)
}
\arguments{
\item{md}{a mudata object}

\item{zipfile}{file to read/write (can also be a directory)}

\item{overwrite}{Pass \code{TRUE} to overwrite if \code{zipfile} already exists.}

\item{expand.tags}{flag to expand tags to columns}

\item{...}{passed to read/write.csv}

\item{validate}{flag to validate mudata object upon read}

\item{load}{a list of csv files (without the .csv extension) to load from the source.}
}
\description{
Read/Write a MuData zip file
}
\examples{
data(longlake2016)
outfile <- tempfile()
write.mudata(longlake2016, outfile)
md <- read.mudata(outfile)
plotgg(subset(md, params=c("Hg", "Pb")))
unlink(outfile)

}

