% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.io.R
\name{write_mudata}
\alias{write_mudata}
\alias{read_mudata}
\alias{write_mudata_zip}
\alias{read_mudata_zip}
\alias{write_mudata_dir}
\alias{read_mudata_dir}
\alias{write_mudata_json}
\alias{to_mudata_json}
\alias{read_mudata_json}
\alias{from_mudata_json}
\title{Read/Write mudata objects}
\usage{
write_mudata(md, filename, ...)

read_mudata(filename, ...)

write_mudata_zip(md, filename, overwrite = FALSE, validate = TRUE,
  update_columns = TRUE, ...)

read_mudata_zip(filename, validate = TRUE, ...)

write_mudata_dir(md, filename, overwrite = FALSE, validate = TRUE,
  update_columns = TRUE, ...)

read_mudata_dir(filename, validate = TRUE, ...)

write_mudata_json(md, filename, overwrite = FALSE, validate = TRUE,
  update_columns = TRUE, pretty = TRUE, ...)

to_mudata_json(md, validate = TRUE, update_columns = TRUE, pretty = FALSE,
  ...)

read_mudata_json(filename, validate = TRUE, ...)

from_mudata_json(txt, validate = TRUE, ...)
}
\arguments{
\item{md}{a mudata object}

\item{filename}{file to read/write (can also be a directory)}

\item{...}{passed to read/write functions}

\item{overwrite}{Pass \code{TRUE} to overwrite if \code{zipfile} already exists.}

\item{validate}{flag to validate mudata object upon read or before write}

\item{update_columns}{Update the columns table "type" column to reflect the internal
R types of columns (reccommended).}

\item{pretty}{Produce pretty or minified JSON output}

\item{txt}{JSON text from which to read a mudata object.}
}
\description{
Read/Write mudata objects
}
\examples{
data(kentvillegreenwood)
# read/write to zip
outfile <- tempfile(fileext=".zip")
#write_mudata(kentvillegreenwood, outfile)
#md <- read_mudata(outfile)
#md <- read_mudata(outfile, retype=TRUE)
unlink(outfile)

# read/write to JSON
outfile <- tempfile(fileext=".json")
#write_mudata(kentvillegreenwood, outfile)
#md <- read_mudata(outfile)
#md <- read_mudata(outfile, retype=TRUE)
unlink(outfile)

}
