% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{as.qtag}
\alias{as.qtag}
\title{Convert an object to a qualifier/tag structure}
\usage{
as.qtag(df, .qualifiers, .values, .tags, quiet = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} or similar object}

\item{.qualifiers}{Column names of qualifying values}

\item{.values}{Column names containing the values of interest}

\item{.tags}{Column names of tag values}

\item{quiet}{Use \code{quiet=TRUE} to suppress error messages}
}
\value{
An object of type \code{qtag}, which is essentially the unchanged
  input with qualifiers, values, and tags information attached.
}
\description{
Convert an object to a qualifier/tag structure
}
\examples{
data("pocmaj")
pocmaj <- as.qtag(pocmaj, qualifiers = c("core", "depth"))
long(pocmaj)
aggregate(pocmaj)
aggregate(long(pocmaj))

}

