% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.mudata}
\alias{autoplot.mudata}
\alias{plot.mudata}
\title{Autoplot a mudata object}
\usage{
\method{autoplot}{mudata}(x, ...)

\method{plot}{mudata}(x, ...)
}
\arguments{
\item{x}{A \link{mudata} object}

\item{...}{Passed on to \link{long_plot} or \link{long_ggplot}}
}
\description{
Produces a quick graphical summary of a mudata object using the ggplot or base plotting
framework.
}
\examples{
data(kentvillegreenwood)
# plot using base plot
plot(kentvillegreenwood)

# a prettier plot using ggplot
library(ggplot2)
autoplot(kentvillegreenwood)

}
