% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.mudata}
\alias{autoplot.mudata}
\alias{plot.mudata}
\title{Autoplot a mudata object}
\usage{
\method{autoplot}{mudata}(object, facets = "param", col = "location",
  pch = "dataset", ...)

\method{plot}{mudata}(x, facets = "param", col = "location",
  pch = "dataset", ...)
}
\arguments{
\item{facets}{Column to be used as facet column}

\item{col}{Column to be used as colour aesthetic}

\item{pch}{Column to be used as shape aesthetic}

\item{...}{Passed on to \link{long_plot} or \link{long_ggplot}}

\item{x, object}{A \link{mudata} object}
}
\description{
Produces a quick graphical summary of a mudata object using the ggplot or base plotting
framework.
}
\examples{
data(kentvillegreenwood)
# plot using base plot
plot(kentvillegreenwood)

# a prettier plot using ggplot
library(ggplot2)
autoplot(kentvillegreenwood)

}
