% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.mudata}
\alias{autoplot.mudata}
\alias{plot.mudata}
\title{Autoplot a mudata object}
\usage{
\method{autoplot}{mudata}(x, ...)

\method{plot}{mudata}(x, ...)
}
\arguments{
\item{x}{A \link{mudata} object}

\item{...}{Passed on to \link{qualifierplot}}
}
\value{
A ggplot object
}
\description{
Produces a quick graphical summary of a mudata object using the ggplot framework.
The \code{subset} argument is quite powerful for filtering, but does not affect the order
of appearance. For this, use \link{subset.mudata}, which by default orders the datasets,
locations, and parameters passed into the function.
}
\examples{
data(kentvillegreenwood)
plot(kentvillegreenwood)
library(ggplot2)
autoplot(kentvillegreenwood)

}
