% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_qtag.R
\name{autoplot.qtag.long}
\alias{autoplot.qtag.long}
\alias{autoplot.qtag.wide}
\alias{plotgg}
\title{Autoplot a qualifier/tag structure}
\usage{
\method{autoplot}{qtag.long}(x, subset, xvar, yvar, facets, ...)

\method{autoplot}{qtag.wide}(x, ...)

plotgg(x, ...)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{subset}{Subset to plot}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{facets}{Column to be used as facetting variable}

\item{...}{Passed on to \code{aes_string()}}
}
\value{
A ggplot object
}
\description{
Autoplot a qualifier/tag structure
}
\examples{
library(ggplot2)
data(pocmaj)
plot.gg(pocmaj)

pocmajqt <- as.qtag(pocmaj, qualifiers=c("core", "depth"))
autoplot(pocmajqt)
autoplot(pocmajqt, subset=core=="MAJ-1" & column \%in\% c("Ca", "Ti"))
autoplot(pocmajqt, shape="core")


}

