% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot.mudata}
\alias{biplot.mudata}
\alias{autobiplot.mudata}
\title{Biplot a mudata object}
\usage{
\method{biplot}{mudata}(x, ...)

\method{autobiplot}{mudata}(x, ...)
}
\arguments{
\item{x}{A mudata object}

\item{...}{passed to plotting methods}
}
\value{
a ggplot object
}
\description{
Uses the ggplot framework and \code{facet_grid} to produce a biplot of a mudata object.
}
\examples{
data(kentvillegreenwood)
kvtemp <- subset(kentvillegreenwood, params = c("mintemp", "maxtemp", "meantemp"))

# use base plotting for regular biplot function
biplot(kvtemp)

# use ggplot and facet_grid to biplot
autobiplot(kvtemp, col = "location")

}
