% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot.mudata}
\alias{biplot.mudata}
\title{Biplot a mudata object}
\usage{
\method{biplot}{mudata}(x, ..., namecolumn = "param")
}
\arguments{
\item{x}{A mudata object}

\item{...}{passed to \link{biplot.qtag.long}}

\item{namecolumn}{The column that contains the names for biplotting}
}
\value{
a ggplot object
}
\description{
Biplot a mudata object
}
\examples{
data(kentvillegreenwood)
biplot(kentvillegreenwood, c("meantemp", "maxtemp", "mintemp"), col="location")

}

