% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{condense.tags}
\alias{condense.tags}
\title{Condense multiple columns to a single JSON column}
\usage{
condense.tags(df, tagcolumns, tagcolumn = "tags")
}
\arguments{
\item{df}{a data frame with tag columns}

\item{tagcolumns}{column names to be condensed to JSON}
}
\value{
A modified data.frame
}
\description{
Condense multiple columns to a single JSON column
}

