% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{group}
\alias{group}
\title{Group a qualifier/tag structure}
\usage{
group(qtag, quiet = FALSE)
}
\arguments{
\item{qtag}{}
}
\value{
A \code{dplyr} grouped data frame
}
\description{
Essentially a shortcut for grouping a \link{qtag} object by its qualifiers
}
\examples{
data(pocmaj)
library(dplyr)
pocmajqt <- as.qtag(pocmaj, qualifiers = c("core", "depth"))
pocmajqt \%>\% group() \%>\% summarise(mean(Ca))
# equivalent to
pocmaj \%>\% group_by(core, depth) \%>\% summarise(mean(Ca))


}

