% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{mudata}
\alias{mudata}
\title{Create a MuData object}
\usage{
mudata(data, locations = NULL, params = NULL, datasets = NULL,
  columns = NULL, x_columns = NULL, dataset_id = "default",
  location_id = "default", validate = TRUE)
}
\arguments{
\item{data}{The data table, which is a molten data frame containing the columns (at least)
'dataset', 'location', 'x', 'param', and 'value'. The 'dataset' column can be omitted
if there is only one dataset contained in the object (its name can be specified by
passing the parameter \code{dataset.id}). The 'location' column can be omitted if
there is only data for one dataset and one location (its name can be specified by
passing the parameter \code{location.id}).}

\item{locations}{The locations table, which is a data frame containing the columns (at least)
'dataset', and 'location'. If omitted, it will be created automatically using all unique
dataset/location combinations.}

\item{params}{The params table, which is a data frame containing the columns (at least)
'dataset', and 'param'. If omitted, it will be created automatically using all unique
dataset/param combinations.}

\item{datasets}{The datasets table, which is a data frame containing the column (at least)
'dataset'. If omitted, it will be generated automatically using all unique datasets.}

\item{columns}{The columns table, which is a data frame containing the columns (at least)
'dataset', 'table', and 'column'. If omitted, it will be created automatically using 
all dataset/table/column combinations.}

\item{x_columns}{A vector of column names from the data table that in combination with
'dataset', 'location', and 'param' identify unique rows.}

\item{dataset_id}{The dataset id to use if a datasets column is omitted.}

\item{location_id}{The location id if a location column is omitted.}

\item{validate}{Pass \code{FALSE} to skip validation of input tables.}
}
\value{
A \code{mudata} object
}
\description{
Create an object describing multi-parameter spatiotemporal data in the (mostly) universal
data format. This format is a collection of tables as described below. For an example
of data already in this format, see the \link{kentvillegreenwood} dataset.
}
\examples{
library(tidyr)
library(dplyr)
data(pocmaj)

# gather columns and summarise replicates
datatable <- pocmaj \%>\%
  gather(Ca, Ti, V, key = "param", value = "param_value") \%>\%
  group_by(core, param, depth) \%>\%
  summarise(value=mean(param_value), sd=mean(param_value)) \%>\%
  rename(location = core)

# create mudata object
mudata(datatable)

}
