% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{mudata}
\alias{mudata}
\title{Create a MuData object}
\usage{
mudata(data, locations = NULL, params = NULL, datasets = NULL,
  dataset.id = "default", location.id = "default", defactorize = TRUE,
  validate = TRUE, expand.tags = TRUE)
}
\arguments{
\item{data}{The data table}

\item{locations}{The locations table (can be omitted)}

\item{params}{The params table (can be omitted)}

\item{datasets}{The datasets table (can be omitted)}

\item{dataset.id}{The dataset id (if datasets is unspecified)}

\item{location.id}{The location id (if locations is unspecified)}

\item{validate}{Flag to validate input}
}
\value{
A \code{mudata} object
}
\description{
Create a MuData object
}

