% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata_app.R
\name{mudata_app}
\alias{mudata_app}
\title{Interactive Plot of a MUData object}
\usage{
mudata_app(obj = NULL)
}
\arguments{
\item{obj}{A mudata object}
}
\description{
Interactive Plot of a MUData object
}
\examples{
if(interactive()) {
  mudata_app()
}

}
