% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{new_mudata}
\alias{new_mudata}
\alias{validate_mudata}
\title{Create, validate a MUData object}
\usage{
new_mudata(md, x_columns)

validate_mudata(md, check_unique = TRUE, check_references = TRUE,
  action = stop)
}
\arguments{
\item{md}{An object of class 'mudata'}

\item{check_unique}{Check if columns identify unique values in the appropriate tables}

\item{check_references}{Check the referential integrity of the mudata object}
}
\description{
Validates a MUData object by calling \code{stop} when an error is found.
}
\examples{
data(kentvillegreenwood)
validate_mudata(kentvillegreenwood)
new_mudata(kentvillegreenwoods)

}
