% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{plotgg.mudata}
\alias{plotgg.mudata}
\title{Autoplot a mudata object}
\usage{
\method{plotgg}{mudata}(x, ...)
}
\arguments{
\item{x}{A \link{mudata} object}

\item{...}{Passed on to \code{plotgg.qtag.long}}
}
\value{
A ggplot object
}
\description{
If you get a \code{seq...finite values} error, you may have to check for params
that have all non-detect values. This can be done with the dplyr summarise function
(\code{group_by(dataset, param) / summarise(allnd=all(is.na(value))) / data.frame()}).
The \code{subset} argument is quite powerful for filtering, but does not affect the order
of appearance. For this, use \link{subset.mudata}.
}

