% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pocmaj}
\alias{pocmaj}
\title{Multi-qualifier paleolimnological concentration data}
\format{A \code{data.frame} containing multi-qualifier concentration data}
\usage{
pocmaj
}
\description{
Base qtag example dataset
}
\keyword{datasets}

