% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pocmajpb210}
\alias{pocmajpb210}
\title{Pre-summarised Sample Data}
\format{A \code{data.frame} containing multi-qualifier data}
\usage{
pocmajpb210
}
\description{
A small example data.frame of pre-summarised data.
}
\keyword{datasets}

