% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pocmajpb210}
\alias{pocmajpb210}
\title{Sample paleolimnological Pb-210 data}
\format{A \code{data.frame} containing multi-qualifier concentration data}
\usage{
pocmajpb210
}
\description{
Example of pre-summarised data
}
\keyword{datasets}

