% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{qualifierplot}
\alias{autoplot.qtag.long}
\alias{autoplot.qtag.wide}
\alias{plot.qtag.long}
\alias{plot.qtag.wide}
\alias{qualifierplot}
\title{Autoplot a qualifier/tag structure}
\usage{
qualifierplot(x, id.vars, values, subset, xvar, yvar, facets, geom = "path",
  errors = "err", ...)

\method{autoplot}{qtag.long}(x, ...)

\method{autoplot}{qtag.wide}(x, ...)

\method{plot}{qtag.long}(x, ...)

\method{plot}{qtag.wide}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{id.vars}{Columns that identify unique values}

\item{values}{Column that contains values to be plotted}

\item{subset}{Subset to plot}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{facets}{Column to be used as facetting variable}

\item{geom}{GGPlot geometries to be used. Can be any combination of point, path, or line.}

\item{errors}{The colum that contains uncertainty information}

\item{...}{Passed on to \code{aes_string()}}
}
\value{
A ggplot object
}
\description{
Autoplot a qualifier/tag structure
}
\examples{
data(pocmaj)
qualifierplot(pocmaj, c("core", "depth"), "Ca")

pocmajqt <- as.qtag(pocmaj, id.vars=c("core", "depth"))
plot(pocmajqt, geom=c("path", "point"))
plot(pocmajqt, subset=core=="MAJ-1" & param \%in\% c("Ca", "Ti"))
plot(pocmajqt, shape="core", geom=c("path", "point"))
plot(long(pocmajqt))

library(ggplot2)
autoplot(pocmajqt, col="core")

}

