% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata2-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{everything}
\alias{matches}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{num_range}
\alias{one_of}
\alias{src_tbls}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{contains}}, \code{\link[dplyr:reexports]{ends_with}}, \code{\link[dplyr:reexports]{everything}}, \code{\link[dplyr:reexports]{matches}}, \code{\link[dplyr:reexports]{num_range}}, \code{\link[dplyr:reexports]{one_of}}, \code{\link[dplyr]{src_tbls}}, \code{\link[dplyr:reexports]{starts_with}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

