% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{everything}
\alias{matches}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{num_range}
\alias{one_of}
\alias{autoplot}
\alias{src_tbls}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{everything}}, \code{\link[dplyr]{matches}}, \code{\link[dplyr]{starts_with}}, \code{\link[dplyr]{ends_with}}, \code{\link[dplyr]{contains}}, \code{\link[dplyr]{num_range}}, \code{\link[dplyr]{one_of}}, \code{\link[dplyr]{src_tbls}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

