% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{second_lake_temp}
\alias{second_lake_temp}
\title{Second Lake Thermistor String Data}
\format{A \link{mudata} object}
\usage{
second_lake_temp
}
\description{
Temperatures at multiple depths in the water column for a season at Second Lake, Lower Sackville,
Nova Scotia, Canada.
}
\examples{
print(second_lake_temp)

}
\references{
Misiuk B (2014) A multi-proxy comparative paleolimnological study of anthropogenic
impact between First and Second Lake, Lower Sackville, Nova Scotia.
B.Sc.H. Thesis, Acadia University. \url{http://scholar.acadiau.ca/islandora/object/theses:1148}
}
\keyword{datasets}
