% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{subset.mudata}
\alias{subset.mudata}
\title{Subset a MuData object}
\usage{
\method{subset}{mudata}(x, datasets = NULL, params = NULL,
  locations = NULL, validate = TRUE, defactorize = FALSE, ...)
}
\arguments{
\item{x}{The object to subset}

\item{datasets}{Vector of datasets to include}

\item{params}{Vector of parameters to include}

\item{locations}{Vector of locations to include}

\item{validate}{Flag to validate output}

\item{defactorize}{Pass \code{FALSE} to keep input columns as factors (may cause errors).}

\item{...}{Aguments to/from methods}
}
\value{
A subsetted MuData object
}
\description{
Subset a MuData object
}

