% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{summary.mudata}
\alias{print.mudata}
\alias{summary.mudata}
\title{Object summary for a mudata object}
\usage{
\method{summary}{mudata}(object, ..., digits = NA)

\method{print}{mudata}(x, ..., digits = 4)
}
\arguments{
\item{object}{A \link{mudata} object}

\item{...}{Unused}

\item{digits}{The number of digits to be used for rounding, or NA to suppress rounding.}

\item{x}{A \link{mudata} object}
}
\value{
A data.frame containing the summary
}
\description{
Returns a data.frame containing summary (by dataset, location, and param)
statistics.
}
\examples{
data(kentvillegreenwood)
summary(kentvillegreenwood, digits=2)
summary(kentvillegreenwood)
print(kentvillegreenwood, digits=2)

}

