% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{wide}
\alias{wide}
\alias{wide.default}
\alias{wide.qtag.long}
\alias{wide.qtag.wide}
\title{Convert data to wide format}
\usage{
wide(x, ...)

\method{wide}{default}(x, ...)

\method{wide}{qtag.wide}(x)

\method{wide}{qtag.long}(qtag, colvar, fun.aggregate, ..., quiet = FALSE)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{...}{Passed to other methods}
}
\value{
A (possibly unchanged) \code{qtag.wide} data.frame
}
\description{
Convert data to wide format
}
\examples{
data(pocmaj)
pocmaj <- as.qtag(pocmaj, qualifiers = c("core", "depth"))
pocmajlong <- long(pocmaj)
wide(pocmajlong)
wide(pocmaj)

}

