% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{write.mudata}
\alias{read.mudata}
\alias{write.mudata}
\title{Read/Write a MuData zip file}
\usage{
write.mudata(md, zipfile, overwrite = FALSE, expand.tags = FALSE, ...)

read.mudata(zipfile, validate = TRUE, expand.tags = TRUE, ...)
}
\arguments{
\item{md}{a mudata object}

\item{zipfile}{file to read/write (can also be a directory)}

\item{expand.tags}{flag to expand tags to columns}

\item{...}{passed to read/write.csv}

\item{validate}{flag to validate mudata object upon read}
}
\description{
Read/Write a MuData zip file
}

